useTranslation moved to hooks
